+++ packages/syft/tests/syft/migrations/protocol_communication_test.py
@@ -1,5 +1,6 @@
 # stdlib
 from copy import deepcopy
+import os
 from pathlib import Path
 from unittest import mock
 
@@ -174,10 +175,11 @@ def my_stage_protocol(protocol_file: Path):
         dp.save_history(dp.protocol_history)
 
         # Cleanup release dir, remove unused released files
-        for _file_path in protocol_release_dir().iterdir():
-            for version in dp.read_json(_file_path):
-                if version not in dp.protocol_history.keys():
-                    _file_path.unlink()
+        if os.path.exists(protocol_release_dir()):
+            for _file_path in protocol_release_dir().iterdir():
+                for version in dp.read_json(_file_path):
+                    if version not in dp.protocol_history.keys():
+                        _file_path.unlink()
 
 
 @pytest.mark.skip(
